Truly independent visualization
Allows to attempt to play the simulation at a given speedup, drops or
reuses frames to achieve the desired speed and reduce the speedup if the
simulation cannot keep up.

Also, blocks the simulation with a blocking buffer if necessary to
preserve data.

Could still be optimised by well designed waits to reduce CPU
consumption instead of reusing frames for the relevant cases.